Greyscale completion menu
This pull request changes the colour of the selected item in the completion menu from a white foreground and a teal background to a simpler, greyscale setup: bold white foreground with medium grey background (note these colour names do not match up with the ones at the top of the file). Originally I made all the colours in the menu one shade darker, mimicking what completion menus often look like in GUI text editors. However, the background was the same as the cursorline, which was quite distracting (as the menu and the cursorline looked ‘connected’). So, the final version only changes the colour of the selected match, as well as both elements of the scroll bar to fit better.